Agentic AI/LangGraph Learning/Krish Sir YouTube/pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.2",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "langsmith>=0.4.1",
    "python-dotenv>=1.1.0",
]

[tool.uv.workspace]
members = ["MCP Server"]
